+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
+export { useGenerateTemplatesMutation } from './useGenerateTemplatesMutation';
 export { useMosaicConfiguration } from './useMosaicConfiguration';
 export { useMosaicConfigurationMutation } from './useMosaicConfigurationMutation';
 export { useOrderTemplates } from './useOrderTemplates';