+++ src/settings/ConfigurationOptions/ConfigurationOptions.tsx
@@ -3,8 +3,10 @@ import React from 'react';
 import { useShowCallout } from '@folio/stripes-acq-components';
 
 import {
+  useGenerateTemplatesMutation,
   useMosaicConfiguration,
   useMosaicConfigurationMutation,
+  useOrderTemplates,
 } from '../../hooks';
 import type { MosaicConfiguration } from '../../types';
 import ConfigurationOptionsForm from './ConfigurationOptionsForm';
@@ -15,9 +17,9 @@ const ConfigurationOptions: React.FC = () => {
   const showCallout = useShowCallout();
 
   const {
-    isLoading,
+    isLoading: isConfigurationLoading,
     mosaicConfiguration,
-    refetch,
+    refetch: refetchConfiguration,
   } = useMosaicConfiguration();
 
   const {
@@ -25,11 +27,22 @@ const ConfigurationOptions: React.FC = () => {
     mutateAsync: mutateMosaicConfiguration,
   } = useMosaicConfigurationMutation();
 
+  const {
+    isLoading: isDefaultTemplateGenerating,
+    mutateAsync: generateDefaultTemplate,
+  } = useGenerateTemplatesMutation();
+
+  const {
+    isLoading: isOrderTemplatesLoading,
+    orderTemplates,
+    refetch: refetchOrderTemplates,
+  } = useOrderTemplates();
+
   const onSubmit = (config: MosaicConfiguration) => {
     return mutateMosaicConfiguration({ config })
       .then(async () => {
         showCallout({ messageId: 'ui-mosaic-settings.sections.configuration-options.submit.success' });
-        await refetch();
+        await refetchConfiguration();
       })
       .catch(() => {
         showCallout({
@@ -39,11 +52,30 @@ const ConfigurationOptions: React.FC = () => {
       });
   };
 
+  const onGenerateTemplates = async () => {
+    generateDefaultTemplate(null)
+      .then(() => {
+        showCallout({ messageId: 'ui-mosaic-settings.sections.configuration-options.action.generate-integration-templates.success' });
+        refetchOrderTemplates();
+      })
+      .catch(() => {
+        showCallout({
+          messageId: 'ui-mosaic-settings.sections.configuration-options.action.generate-integration-templates.error',
+          type: 'error',
+        });
+      });
+  };
+
+  const isLoading = isConfigurationLoading || isMutationLoading || isDefaultTemplateGenerating;
+
   return (
     <ConfigurationOptionsForm
-      onSubmit={onSubmit}
       initialValues={mosaicConfiguration || DEFAULT_INITIAL_VALUES}
-      isLoading={isLoading || isMutationLoading}
+      isLoading={isLoading}
+      isOrderTemplatesLoading={isOrderTemplatesLoading}
+      onGenerateTemplates={onGenerateTemplates}
+      onSubmit={onSubmit}
+      orderTemplates={orderTemplates}
     />
   );
 };