+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.0 (IN PROGRESS)
 
 * [UIMS-6](https://folio-org.atlassian.net/browse/UIMS-6) Remove incorrect dropdown reference from ui-mosaic-settings description.
+* [UIMS-7](https://folio-org.atlassian.net/browse/UIMS-7) Add button to the configuration options settings for default templates generating.
 
 ## [1.0.0](https://github.com/folio-org/ui-mosaic-settings/tree/v1.0.0) (2025-09-22)